requirements.txt
zc.buildout
Pillow==6.2.2
plonecli
tox 
